requirements.txt
streamlit
openai-whisper
torch
transformers
accelerate
